Add SamplerDesc struct for texture sampling parameters
Introduces a new record struct `SamplerDesc` in the `Zenith.NET` namespace that implements the `IDesc` interface. This struct includes properties for U, V, and W address modes, filter type, comparison function, maximum anisotropy, and level of detail settings. A `Validate` method is also added to ensure the integrity of the level of detail values.