Add VKValidationLayer and improve VKGraphicsContext cleanup
Introduce VKValidationLayer to handle Vulkan validation layers
and integrate it into VKGraphicsContext. Update the constructor
of VKGraphicsContext to initialize the validation layer based
on the useValidationLayer parameter. Add a Destroy method to
VKGraphicsContext for proper Vulkan resource cleanup, including
waiting for the device to become idle and destroying Vulkan
instances and devices. These changes enhance resource management
and extend functionality.