Simplify memory access in VKRayTracingPipeline loop
Refactor the memory access logic in the `VKRayTracingPipeline` class
to eliminate the use of a separate `mappedPointer` variable. The
memory offset is now directly calculated within the `Unsafe.CopyBlock`
call as `(byte*)mappedMemory.Pointer + (HandleSizeAligned * i)`.
This change improves code clarity and reduces the number of variables
used in the loop.